@@ -1,74 +1,44 @@
+[project]
+name = "EVE_TEST_BOT"
+version = "0.1.0"
+description = "EVE_TEST_BOT"
+readme = "README.md"
+requires-python = ">=3.8, <4.0"
+
+[tool.nonebot]
+adapters = [
+    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
+]
+plugins = ["nonebot_plugin_alconna", "nonebot_plugin_htmlrender", "nonebot_plugin_apscheduler"]
+plugin_dirs = ["src/plugins"]
+builtin_plugins = []
+
 [tool.poetry]
-name = "nonebot_plugin_eve_tool"
+name = "nonebot-plugin-eve-tool"
 version = "0.1.0"
 description = "A NoneBot based plugin provides information query services for EVE online."
 authors = ["zifox666 <zifox666@gmail.com>"]
-repository = "https://github.com/zifox666/nonebot-plugin-eve-tool"
+license = "GPL-3.0 license"
+readme = "README.md"
 
 [tool.poetry.dependencies]
-python = "^3.9"
+python = "^3.10"
+aioredis = "^2.0.1"
+aiomysql = "^0.2.0"
+aiohttp = "^3.9.5"
+jieba = "^0.42.1"
+nonebot-plugin-alconna = "^0.47.1"
+requests = "^2.32.3"
+tqdm = "^4.66.4"
+qrcode = "^7.4.2"
+matplotlib = "^3.9.0"
+numpy = "^2.0.0"
+scipy = "^1.13.1"
 
-[tool.poetry.dependencies]
-python = "^3.9"
-aiohttp = "3.9.5"
-aiomysql = "0.2.0"
-aioredis = "2.0.1"
-aiosignal = "1.3.1"
-annotated-types = "0.6.0"
-anyio = "4.3.0"
-async-timeout = "4.0.3"
-attrs = "23.2.0"
-certifi = "2024.2.2"
-charset-normalizer = "3.3.2"
-click = "8.1.7"
-colorama = "0.4.6"
-exceptiongroup = "1.2.1"
-fastapi = "0.110.2"
-frozenlist = "1.4.1"
-h11 = "0.14.0"
-httpcore = "1.0.5"
-httptools = "0.6.1"
-httpx = "0.27.0"
-idna = "3.7"
-importlib_metadata = "7.1.0"
-linkify-it-py = "2.0.3"
-loguru = "0.7.2"
-markdown-it-py = "3.0.0"
-mdit-py-plugins = "0.4.0"
-mdurl = "0.1.2"
-msgpack = "1.0.8"
-multidict = "6.0.5"
-nonebot-adapter-onebot = "2.4.3"
-nonebot2 = "2.2.1"
-nonechat = "0.2.1"
-pydantic = "2.7.1"
-pydantic-settings = "2.2.1"
-pydantic_core = "2.18.2"
-Pygments = "2.17.2"
-pygtrie = "2.5.0"
-PyMySQL = "1.1.1"
-pypng = "0.20220715.0"
-python-dotenv = "1.0.1"
-PyYAML = "6.0.1"
-qrcode = "7.4.2"
-requests = "2.32.3"
-rich = "13.7.1"
-sniffio = "1.3.1"
-starlette = "0.37.2"
-tenacity = "8.2.3"
-textual = "0.29.0"
-tomli = "2.0.1"
-tqdm = "4.66.4"
-typing_extensions = "4.11.0"
-uc-micro-py = "1.0.3"
-urllib3 = "2.2.2"
-uvicorn = "0.29.0"
-watchfiles = "0.21.0"
-websockets = "12.0"
-win32-setctime = "1.1.0"
-yarl = "1.9.4"
-zipp = "3.18.1"
+[[tool.poetry.source]]
+name = "aliyun"
+url = "https://mirrors.aliyun.com/pypi/simple"
 
 [build-system]
-requires = ["poetry-core>=1.0.0"]
+requires = ["poetry-core"]
 build-backend = "poetry.core.masonry.api"
