@@ -1,5 +1,7 @@
 import json
 
+from nonebot import logger
+
 from .RedisArray import RedisArray
 from ...model.config import plugin_config
 from ...utils.common import pack_strings
@@ -14,13 +16,12 @@ async def get_names_for_redis(ids: int | str) -> str | list[str]:
     :return:
     """
     names_list = []
-    if isinstance(ids, int):
+    if isinstance(ids, str | int):
         ids = [ids]
     for _id in ids:
         name = await RA.hget(f'eve_type:{_id}',
                              'name' if plugin_config.eve_lagrange_preference == 'zh' else 'name_en')
-        names_list.append(pack_strings(name))
-    print(names_list)
+        names_list.append(name)
     return names_list[0] if len(names_list) == 1 else names_list
 
 
@@ -37,12 +38,13 @@ async def get_alias_for_redis(alias: str) -> list | None:
     if not data:
         return None
     else:
-        print(f"查询到物品别名:\n{data}")
+        logger.info(f"查询到物品别名:\n{data}")
         return data
 
 
-async def search_eve_types(query_list, lang='zh', fuzzy: bool = True):
+async def search_eve_types_for_redis(query_list, lang='zh', fuzzy: bool = True):
     """
+    !!!废弃!!!
     在 Redis 中根据指定语言进行模糊搜索
     :param query_list: 查询字符串或字符串列表
     :param lang: 语言类型 ('zh' 表示中文, 'en' 表示英文)
@@ -56,7 +58,7 @@ async def search_eve_types(query_list, lang='zh', fuzzy: bool = True):
     total_count = 0
 
     for query in query_list:
-        search_query = f'@name:{" ".join(query)}' if lang == 'zh' else f'@name_en:{query}*'
+        search_query = f'@name:{query}' if lang == 'zh' else f'@name_en:{query}*'
 
         result = await RA.execute('FT.SEARCH', 'eveTypeIdx', search_query, 'RETURN', '2', 'name',
                                   'name_en', 'LIMIT', '0', '50')
@@ -75,3 +77,13 @@ async def extract_ids_from_results(results):
         items_list.append(eve_type_id)
     items_list.sort()
     return results[0], items_list
+
+
+async def get_price_from_cache(ids: list[int]) -> dict:
+    result = {}
+    for _id in ids:
+        result[str(_id)] = json.loads(await RA.hget('market_price', f"@eve_type:{_id}"))
+    if result:
+        return result
+
+
