@@ -19,7 +19,7 @@ async def load_sde_to_redis(RA: RedisArray, MYSQL: MysqlArray) -> bool:
     for row in eve_type_data:
         redis_key = f"eve_type:{row['id']}"
         if row['name']:
-            await RA.hset(redis_key, 'name', ' '.join(row['name']))
+            await RA.hset(redis_key, 'name', row['name'])
         if row['name_en']:
             await RA.hset(redis_key, 'name_en', row['name_en'])
 
