@@ -10,17 +10,16 @@ async def process_orders(data: List[dict], type_id: int) -> dict:
     """
     min_sell_price = float('inf')
     max_buy_price = float('-inf')
-    min_sell_volume = 0
-    max_buy_volume = 0
-
+    sell_volume = 0
+    buy_volume = 0
     for order in data:
         if not order['is_buy_order']:
-            min_sell_volume += order['volume_remain']
+            sell_volume += order['volume_remain']
             if order['price'] < min_sell_price:
                 min_sell_price = order['price']
 
         else:
-            max_buy_volume += order['volume_remain']
+            buy_volume += order['volume_remain']
             if order['price'] > max_buy_price:
                 max_buy_price = order['price']
 
@@ -31,8 +30,35 @@ async def process_orders(data: List[dict], type_id: int) -> dict:
 
     return {
         'sell': min_sell_price,
-        'sell_num': min_sell_volume,
+        'sell_num': sell_volume,
         'buy': max_buy_price,
-        'buy_num': max_buy_volume,
+        'buy_num': buy_volume,
         'name': await search.get_names_for_redis(type_id)
     }
+
+
+async def process_all(all_data: List[dict]) -> dict:
+    price_list = {}
+    min_sell_price = float('inf')
+    max_buy_price = float('-inf')
+    sell_volume = 0
+    buy_volume = 0
+    for order in all_data:
+        if order["type_id"] not in price_list:
+            price_list[order["type_id"]] = {
+                'sell': min_sell_price,
+                'sell_num': sell_volume,
+                'buy': max_buy_price,
+                'buy_num': buy_volume,
+                'name': await search.get_names_for_redis(order["type_id"])
+            }
+        if not order['is_buy_order']:
+            price_list[order["type_id"]]["sell_num"] += order['volume_remain']
+            if order['price'] < price_list[order["type_id"]]["sell"]:
+                price_list[order["type_id"]]["sell"] = order['price']
+        else:
+            price_list[order["type_id"]]["buy_num"] += order['volume_remain']
+            if order['price'] > price_list[order["type_id"]]["buy"]:
+                price_list[order["type_id"]]["buy"] = order['price']
+    return price_list
+
